# lbaw23113_openapi.yaml
openapi: 3.0.0
info:
 version: '1.0'
 title: 'LBAW LeiloArte Web API'
 description: 'Web Resources Specification (A7) for LeiloArte'
servers:
- url: http://lbaw.fe.up.pt
  description: Production server
...
paths:
...
 /users/{id}/auctions:
    get:
      operationId: R110
      summary: 'R110: User Auctions'
      description: 'Display user auctions. Access: USR, ADM'
      tags:
        - 'M01: Authentication and Individual Profile'
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Display user auctions UI
        '404':
          description: User not found
...
 /auctions/{id}:
    get:
      operationId: R203
      summary: 'R203: Auction'
      description: 'Display auction. Access: PUB, ADM'
      tags:
        - 'M04: Auctions'
      parameters:
        - name: id
          in: path
          description: Auction ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Display auction UI
        '404':
          description: Auction not found{
...